% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{any_na}
\alias{any_na}
\title{function that checks if any element is na in a vector}
\usage{
any_na(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical value if there is at least one \code{NA} value
}
\description{
function that checks if any element is na in a vector
}
