% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcetracker_functions.R
\name{aux_get_matrix}
\alias{aux_get_matrix}
\title{gets the ASV matrix for some samples}
\usage{
aux_get_matrix(asv, samples)
}
\arguments{
\item{asv}{the full asv matrix}

\item{samples}{a vector of samples to select}
}
\description{
gets the ASV matrix for some samples
}
