% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcetracker_functions.R
\name{aux_split_run}
\alias{aux_split_run}
\title{splits the ASV matrix according to the negative control structure}
\usage{
aux_split_run(group, asv, split_size)
}
\arguments{
\item{group}{negative control structure indicating the samples for each
group}

\item{asv}{ASV matrix, nsamples x nASVs}

\item{split_size}{maximum split size when partitioning the matrix}
}
\description{
splits the ASV matrix according to the negative control structure
}
