% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{compute_prevalence}
\alias{compute_prevalence}
\title{computes the prevalence statistics for each ASV}
\usage{
compute_prevalence(asv_table, idtaxa)
}
\arguments{
\item{asv_table}{matrix with samples as rows and ASVs as columns}

\item{idtaxa}{matrix with taxa classification for each ASV}
}
\value{
a \code{tibble} with the prevalence statistics per ASV
}
\description{
computes the prevalence statistics for each ASV
}
