% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_error_rates_functions.R
\name{condor_error_rates}
\alias{condor_error_rates}
\title{generate the input to learn the error rates with condor}
\usage{
condor_error_rates(error_queue_files = c("", ""),
  error_prefixes = c("R1", "R2"), error_param_file = "",
  error_outdir = ".", condor_file = "./condor_error_rates",
  batch_name = "dada2_error_rates", request_cores = 4,
  request_mem = "4 GB")
}
\arguments{
\item{error_queue_files}{vector with names of the file with the c(R1,R2) fastq files}

\item{error_prefixes}{vector with the names of the output error rates prefix for the c(R1,R2) files}

\item{error_param_file}{name of the file with the learn error rates parameters}

\item{error_outdir}{directory where the output is going to be saved}

\item{condor_file}{name of the file with condor instructions}

\item{batch_name}{string with the name of the batch}

\item{request_cores}{number of cpus per machine}

\item{request_mem}{number of GB required as memory}
}
\description{
generate the input to learn the error rates with condor
}
