% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_taxa.R
\name{condor_extract_fasta}
\alias{condor_extract_fasta}
\title{generates FASTA files from the ASV matrix}
\usage{
condor_extract_fasta(
  asv_file = "",
  fasta_prefix = "taxa",
  fasta_outdir = ".",
  condor_file = "condor_extract_fasta",
  batch_name = "dada2_extract_fasta",
  request_cores = 4,
  request_mem = "4 GB"
)
}
\arguments{
\item{asv_file}{rds file with the ASV table without chimeras}

\item{fasta_prefix}{name of the output file}

\item{fasta_outdir}{directory where the output is going to be saved}

\item{condor_file}{name of the file with condor instructions}

\item{batch_name}{string with the name of the batch}

\item{request_cores}{number of cpus per machine}

\item{request_mem}{number of GB required as memory}
}
\description{
generates FASTA files from the ASV matrix
}
