% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{condor_generate_all}
\alias{condor_generate_all}
\title{generates all condor files}
\usage{
condor_generate_all(condordir, prefix)
}
\arguments{
\item{condordir}{directory where all condor files are going to be saved}

\item{prefix}{prefix to be used in all files}
}
\description{
generates all condor files
}
