% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{condor_generate_group_dag}
\alias{condor_generate_group_dag}
\title{generate the dagman file for running the pipeline with condor}
\usage{
condor_generate_group_dag(
  condordir,
  prefix,
  filter_trim_file,
  error_rates_file,
  merge_pairs_file,
  seque_table_file
)
}
\arguments{
\item{condordir}{directory where all condor files are going to be saved}

\item{prefix}{prefix to be used in all files}

\item{filter_trim_file}{name of the filter and trim condor file}

\item{error_rates_file}{name of the learning error rates condor file}

\item{merge_pairs_file}{name of the merge pairs condor file}

\item{seque_table_file}{name of the make sequence table condor file}
}
\description{
generate the dagman file for running the pipeline with condor
}
