% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pairs_funs.R
\name{condor_merge_pairs}
\alias{condor_merge_pairs}
\title{generate the input to filter and trim the fastq files
with condor}
\usage{
condor_merge_pairs(merge_queue_file = "", merge_param_file = "",
  merge_outdir = ".", error_rate_files = c("R1.rds", "R2.rds"),
  condor_file = "./condor_merge_pairs",
  batch_name = "dada2_merge_pairs", request_cores = 4)
}
\arguments{
\item{merge_queue_file}{name of the file with the queue, it has 3 columns sample_name | R1.fastq | R2.fastq}

\item{merge_param_file}{name of the file with the merge pairs parameters}

\item{merge_outdir}{directory where the output is going to be saved}

\item{error_rate_files}{vector with names of the learned error rates}

\item{condor_file}{name of the file with condor instructions}

\item{batch_name}{string with the name of the batch}

\item{request_cores}{number of cpus per machine}
}
\description{
generate the input to filter and trim the fastq files
with condor
}
