% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{create_file_structure}
\alias{create_file_structure}
\title{create the full file structure that integrated across different groups}
\usage{
create_file_structure(outdir, groups)
}
\arguments{
\item{outdir}{a string with the name of the base output directoryn}

\item{groups}{a string vector with the names of samples groups}
}
\description{
create the full file structure that integrated across different groups
}
\examples{
\dontrun{
 groups <- c("group1", "group2")
 create_file_structure(outdir, groups)
}
}
