% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{create_file_structure_group}
\alias{create_file_structure_group}
\title{creates the file structure for running dada2 with condor}
\usage{
create_file_structure_group(outdir)
}
\arguments{
\item{outdir}{a string with the name of the base output directoryn}
}
\description{
creates the file structure for running dada2 with condor
}
