% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{dada2_param_copy_all}
\alias{dada2_param_copy_all}
\title{copies all the parameter templates to a directory, and
uses a prefix to identify all}
\usage{
dada2_param_copy_all(outdir, prefix)
}
\arguments{
\item{outdir}{directory where all the parameters are saved}

\item{prefix}{prefix for all the files}
}
\description{
copies all the parameter templates to a directory, and
uses a prefix to identify all
}
