% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{dada2_param_filter_trim}
\alias{dada2_param_filter_trim}
\title{creates a default param file for the filter and trim process}
\usage{
dada2_param_filter_trim(paramfile)
}
\arguments{
\item{paramfile}{name of the output json file with the filter and trim parameters}
}
\description{
creates a default param file for the filter and trim process
}
