% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{dada2_param_learn_error_rates}
\alias{dada2_param_learn_error_rates}
\title{creates a default param file for the learn error rates parameter}
\usage{
dada2_param_learn_error_rates(paramfile)
}
\arguments{
\item{paramfile}{name of the output json file
with the learn error rates parameters}
}
\description{
creates a default param file for the learn error rates parameter
}
