% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{dada2_param_merge_pairs}
\alias{dada2_param_merge_pairs}
\title{creates a default param file for the merge pairs process}
\usage{
dada2_param_merge_pairs(paramfile)
}
\arguments{
\item{paramfile}{name of the output json file with the merge pairs parameters}
}
\description{
creates a default param file for the merge pairs process
}
