% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{dada2_param_remove_chimera}
\alias{dada2_param_remove_chimera}
\title{creates a default param file for the remove chimera process}
\usage{
dada2_param_remove_chimera(paramfile)
}
\arguments{
\item{paramfile}{name of the output json file with the
remove chimera parameters}
}
\description{
creates a default param file for the remove chimera process
}
