% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seqtab_funs.R
\name{get_param_chimeras}
\alias{get_param_chimeras}
\title{function to add default parameters to remove_chimeras script}
\usage{
get_param_chimeras(param_name, param_frame)
}
\arguments{
\item{param_name}{name of the parameter to look for}

\item{param_frame}{data.frame with the chosen parameters}
}
\value{
the option in param_frame of the default value for
    \code{param_name}
}
\description{
function to add default parameters to remove_chimeras script
}
