% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_error_rates_functions.R
\name{get_param_error_rates}
\alias{get_param_error_rates}
\title{function to add default parameters for learn error rates script}
\usage{
get_param_error_rates(param_name, param_frame)
}
\arguments{
\item{param_name}{name of the parameter to look for}

\item{param_frame}{data.frame with the chosen parameters}
}
\value{
the option in param_frame of the default value for
    \code{param_name}
}
\description{
function to add default parameters for learn error rates script
}
