% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_trim_functions.R
\name{get_param_filter_trim}
\alias{get_param_filter_trim}
\title{function to add default parameters for filter and trim script}
\usage{
get_param_filter_trim(param_name, param_frame)
}
\arguments{
\item{param_name}{name of the parameter to look for}

\item{param_frame}{data.frame with the chosen parameters}
}
\value{
the option in param_frame of the default value for
    \code{param_name}
}
\description{
function to add default parameters for filter and trim script
}
