% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pairs_funs.R
\name{get_param_merge_pairs}
\alias{get_param_merge_pairs}
\title{function to add default parameters for merge_pairs script}
\usage{
get_param_merge_pairs(param_name, param_frame)
}
\arguments{
\item{param_name}{name of the parameter to look for}

\item{param_frame}{data.frame with the chosen parameters}
}
\value{
the option in param_frame of the default value for
    \code{param_name}
}
\description{
function to add default parameters for merge_pairs script
}
