% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seqtab_functions.R
\name{get_tax_ids}
\alias{get_tax_ids}
\title{auxiliar function to label the taxa}
\usage{
get_tax_ids(dna, rdata_file, confidence_thr = 60, cores = NULL)
}
\arguments{
\item{dna}{\code{DNAStringSet} with the ASV sequences}

\item{rdata_file}{idtaxa model learned saved as a RData file}

\item{confidence_thr}{confidence threshold, higher levels
indicates higher accuracy}

\item{cores}{number of cpus to use}
}
\value{
a \code{data.frame} with the labelled taxa
}
\description{
auxiliar function to label the taxa
}
