% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcetracker_functions.R
\name{load_sourcetracker}
\alias{load_sourcetracker}
\title{loads sourcetracker code from the raw github repository}
\usage{
load_sourcetracker()
}
\description{
loads sourcetracker code from the raw github repository
}
