% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcetracker_functions.R
\name{merge_sourcetracker_results}
\alias{merge_sourcetracker_results}
\title{merge sourcetracker results into a big sourcetracker object}
\usage{
merge_sourcetracker_results(split_results)
}
\arguments{
\item{split_results}{a list of sourcetracker objects}
}
\value{
a sourcetracker object
}
\description{
merge sourcetracker results into a big sourcetracker object
}
