% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{parse_filtered_file}
\alias{parse_filtered_file}
\title{converts from input files into the filtered file format for dada2 pipeline}
\usage{
parse_filtered_file(fastq, outdir)
}
\arguments{
\item{fastq}{a character vector with the name of a file}

\item{outdir}{a character with the name of the output directory}
}
\value{
a character vector with modified
}
\description{
converts from input files into the filtered file format for dada2 pipeline
}
