% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_taxa.R
\name{parse_kraken_taxa_from_id}
\alias{parse_kraken_taxa_from_id}
\title{parses the taxa obtained by kraken from the ids}
\usage{
parse_kraken_taxa_from_id(kraken_out)
}
\arguments{
\item{kraken_out}{results obtained by kraken with `.out` suffix}
}
\value{
a \code{tibble::tibble} with text ids parsed from the ncbi db
}
\description{
parses the taxa obtained by kraken from the ids
}
