% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_diagnostics.R
\name{plot_abundance_per_step}
\alias{plot_abundance_per_step}
\title{compares the (relative) abundance for each step}
\usage{
plot_abundance_per_step(nreads_summary, summary_fun = median, relative = FALSE)
}
\arguments{
\item{nreads_summary}{output of the \code{summarize_number_reads} function}

\item{summary_fun}{function used to summarize the trend,
the default function is the \code{median}}

\item{relative}{boolean indicator determining wheter comparing the
relative abundance}
}
\value{
a \code{ggplot} object
}
\description{
compares the (relative) abundance for each step
}
