% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{plots the prevalence statistic}
\usage{
plot_prevalence(prevalence, nsamples)
}
\arguments{
\item{prevalence}{tibble with prevalence statistic
computed by \code{compute_prevalence}}

\item{nsamples}{integer with the number of samples in the data}
}
\value{
a \code{ggplot} object with the plot
}
\description{
plots the prevalence statistic
}
