% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_diagnostics.R
\name{plot_quality_profiles}
\alias{plot_quality_profiles}
\title{plots the quality profiles generated by dada2's
\code{plotQualityProfile} function, both ends per page, and all samples in a
single pdf file}
\usage{
plot_quality_profiles(sample_tibble, pdf_filename)
}
\arguments{
\item{sample_tibble}{a tibble / data.frame with 3 columns:
sample_name | end 1 file | end 2 file}

\item{pdf_filename}{a string with the name of the file where the plots are
saved}
}
\description{
plots the quality profiles generated by dada2's
\code{plotQualityProfile} function, both ends per page, and all samples in a
single pdf file
}
