% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance.R
\name{relative_abundance_heatmap}
\alias{relative_abundance_heatmap}
\title{plots a relative abundance heatmap with samples as columns}
\usage{
relative_abundance_heatmap(
  taxa_summary,
  taxa_rank,
  meta,
  meta_key = "sample",
  min_prop = 0.01
)
}
\arguments{
\item{taxa_summary}{a tibble with the abundance and prevalence for each taxa}

\item{taxa_rank}{a string with the taxonomic rank for which the data
is summarized}

\item{meta}{a tibble with the meta information, must contains a key column
with the same value than taxa_summary}

\item{meta_key}{name of that key, by default uses the value "sample"}

\item{min_prop}{the minimum frequency required to group taxa into "Other"}
}
\description{
plots a relative abundance heatmap with samples as columns
}
