% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_files.R
\name{sourcetracker_parameters}
\alias{sourcetracker_parameters}
\title{creates a default param file for the source tracker check}
\usage{
sourcetracker_parameters(paramfile)
}
\arguments{
\item{paramfile}{name of the output json file with the
sourcetracker parameters}
}
\description{
creates a default param file for the source tracker check
}
