% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_diagnostics.R
\name{summarize_number_reads}
\alias{summarize_number_reads}
\title{summarizes the number of reads resulted after each step into one table}
\usage{
summarize_number_reads(reads_file, outprefix, outdir, cores)
}
\arguments{
\item{reads_file, }{a 3 column csv file with columns: sample | R1.fastq | R2.fastq}

\item{outprefix}{prefix used in the final ASV table}

\item{outdir}{names of the output directory where the results are saved}

\item{cores}{number of cpus used to run the process}
}
\description{
summarizes the number of reads resulted after each step into one table
}
