% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance.R
\name{summarize_per_taxa}
\alias{summarize_per_taxa}
\title{summarizes the asv table (in tibble format) grouping by a taxonomy level}
\usage{
summarize_per_taxa(asv_tibble, taxa_rank)
}
\arguments{
\item{asv_tibble}{a tibble with asv information merged with taxonomy
annotations}

\item{taxa_rank}{a string with the taxonomic rank for which the data
is summarized}
}
\value{
a tibble with the abundance and prevalence for each taxa
}
\description{
summarizes the asv table (in tibble format) grouping by a taxonomy level
}
