% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{warning_file}
\alias{warning_file}
\title{creates a warning if the file doesn't exists}
\usage{
warning_file(file, warn_text)
}
\arguments{
\item{file}{a string with the name of the file}

\item{warn_text}{a string with the text to be added as warning}
}
\description{
creates a warning if the file doesn't exists
}
