% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seqtab_funs.R
\name{condor_label_taxa}
\alias{condor_label_taxa}
\title{generate the input file to label taxa with condor}
\usage{
condor_label_taxa(asv_file, taxa_model_file, taxa_prefix, taxa_outdir,
  condor_file, batch_name, request_cores)
}
\arguments{
\item{asv_file}{rds file with the ASV table without chimeras}

\item{taxa_model_file}{RData file with ta learned taxa model to use with \code{DECIPHER}}

\item{taxa_prefix}{name of the output file}

\item{taxa_outdir}{directory where the output is going to be saved}

\item{condor_file}{name of the file with condor instructions}

\item{batch_name}{string with the name of the batch}

\item{request_cores}{number of cpus per machine}
}
\description{
generate the input file to label taxa with condor
}
