% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seqtab_functions.R
\name{condor_remove_chimeras}
\alias{condor_remove_chimeras}
\title{generate the input file to remove chimeras from the sequence table with condor}
\usage{
condor_remove_chimeras(sequence_table_file = "", chim_param_file = "",
  chim_prefix = "remove_chimeras", chim_outdir = ".",
  condor_file = "./condor_remove_chimeras",
  batch_name = "dada2_remove_chimeras", request_cores = 4,
  request_mem = "4 GB")
}
\arguments{
\item{sequence_table_file}{name of the file with the sequence tables to merge (one p/line)}

\item{chim_param_file}{name of the remove chimeras parameter file}

\item{chim_prefix}{prefix name for the sequence table output file}

\item{chim_outdir}{directory where the output is going to be saved}

\item{condor_file}{name of the file with condor instructions}

\item{batch_name}{string with the name of the batch}

\item{request_cores}{number of cpus per machine}

\item{request_mem}{number of GB required as memory}
}
\description{
generate the input file to remove chimeras from the sequence table with condor
}
