% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{create_file_structure}
\alias{create_file_structure}
\title{creates the file structure for running dada2 with condor}
\usage{
create_file_structure(outdir)
}
\arguments{
\item{outdir}{Name of the base output directory}
}
\description{
creates the file structure for running dada2 with condor
}
