% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{save_files}
\alias{save_files}
\title{write csv files structured to make the dada2 pipeline
run with condor}
\usage{
save_files(all_samples, outfile)
}
\arguments{
\item{all_samples}{a \code{tibble} with 3 columns:
name, end1, end2}

\item{outfile}{the file where the full tibble is going to be saved}
}
\description{
write csv files structured to make the dada2 pipeline
run with condor
}
